중복 레이아웃 공통으로 빼기
